+++ packages/sandbox/src/index.ts
@@ -6,6 +6,7 @@ import {
   createURLQueryWithCompilerOptions,
 } from "./compilerOptions"
 import lzstring from "./vendor/lzstring.min"
+
 import { supportedReleases } from "./release_data"
 import { getInitialCode } from "./getInitialCode"
 import { extractTwoSlashCompilerOptions, twoslashCompletions } from "./twoslashSupport"
@@ -113,10 +114,10 @@ export const createTypeScriptSandbox = (
   monaco: Monaco,
   ts: typeof import("typescript")
 ) => {
-  const config = { ...defaultPlaygroundSettings(), ...partialConfig }
-  if (!("domID" in config) && !("elementToAppend" in config))
+  if (!("domID" in partialConfig) && !("elementToAppend" in partialConfig))
     throw new Error("You did not provide a domID or elementToAppend")
 
+  const config = { ...defaultPlaygroundSettings(), ...partialConfig }
   const defaultText = config.suppressAutomaticallyGettingDefaultText
     ? config.text
     : getInitialCode(config.text, document.location)
@@ -144,7 +145,10 @@ export const createTypeScriptSandbox = (
 
   const language = languageType(config)
   const filePath = createFileUri(config, compilerOptions, monaco)
-  const element = "domID" in config ? document.getElementById(config.domID) : (config as any).elementToAppend
+  const element = "elementToAppend" in config ? config.elementToAppend : document.getElementById(config.domID)
+  if (!element)
+    throw new Error("DOM element lookup by domID failed")
+
 
   const model = monaco.editor.createModel(defaultText, language, filePath)
   monaco.editor.defineTheme("sandbox", sandboxTheme)
@@ -291,10 +295,13 @@ export const createTypeScriptSandbox = (
   }
 
   /** Gets the results of compiling your editor's code */
-  const getEmitResult = async () => {
+  const getEmitResult = async (
+    emitOnlyDtsFiles?: boolean,
+    forceDtsEmit?: boolean
+  ) => {
     const model = editor.getModel()!
     const client = await getWorkerProcess()
-    return await client.getEmitOutput(model.uri.toString())
+    return await client.getEmitOutput(model.uri.toString(), emitOnlyDtsFiles, forceDtsEmit)
   }
 
   /** Gets the JS  of compiling your editor's code */
@@ -311,11 +318,14 @@ export const createTypeScriptSandbox = (
     return (firstJS && firstJS.text) || ""
   }
 
+  const isDtsFile = (name: string) => /\.d\.([^\.]+\.)?[cm]?ts$/i.test(name)
+
   /** Gets the DTS for the JS/TS  of compiling your editor's code */
   const getDTSForCode = async () => {
-    const result = await getEmitResult()
-    return result.outputFiles.find((o: any) => o.name.endsWith(".d.ts"))!.text
+    const result = await getEmitResult(/*emitOnlyDtsFiles*/ undefined, /*forceDtsEmit*/ true)
+    return result.outputFiles.find((o: any) => isDtsFile(o.name))?.text || ""
   }
+
 
   const getWorkerProcess = async (): Promise<TypeScriptWorker> => {
     const worker = await getWorker()
@@ -325,7 +335,7 @@ export const createTypeScriptSandbox = (
 
   const getDomNode = () => editor.getDomNode()!
   const getModel = () => editor.getModel()!
-  const getText = () => getModel().getValue()
+  const getText = () => getModel().getValue() || ""
   const setText = (text: string) => getModel().setValue(text)
 
   const setupTSVFS = async (fsMapAdditions?: Map<string, string>) => {