+++ packages/typescriptlang-org/scripts/updateAppInsightsGitHubIssue.js
@@ -7,28 +7,28 @@
 //
 
 const {
-  makeMarkdownOfWeeklyAppInsightsInfo,
-} = require("./makeMarkdownForAppInsights")
-const Octokit = require("@octokit/rest")
-
-// Get this from OneNote
-if (!process.env.GITHUB_TOKEN)
-  throw new Error("No GitHub Token at process.env.GITHUB_TOKEN")
-
-const go = async () => {
-  const octokit = new Octokit({
-    auth: process.env.GITHUB_TOKEN,
-    userAgent: "TS AppInsights Issue Updater",
-  })
-
-  const md = await makeMarkdownOfWeeklyAppInsightsInfo()
-
-  await octokit.issues.update({
-    owner: "Microsoft",
-    repo: "TypeScript-Website",
-    issue_number: 1014,
-    body: md,
-  })
-}
-
-go()
+    makeMarkdownOfWeeklyAppInsightsInfo,
+  } = require("./makeMarkdownForAppInsights")
+  const Octokit = require("@octokit/rest")
+  
+  // Get this from OneNote
+  if (!process.env.GITHUB_TOKEN)
+    throw new Error("No GitHub Token at process.env.GITHUB_TOKEN")
+  
+  const go = async () => {
+    const octokit = new Octokit({
+      auth: process.env.GITHUB_TOKEN,
+      userAgent: "TS AppInsights Issue Updater",
+    })
+  
+    const md = await makeMarkdownOfWeeklyAppInsightsInfo()
+  
+    await octokit.issues.update({
+      owner: "Microsoft",
+      repo: "TypeScript-Website",
+      issue_number: 1014,
+      body: md,
+    })
+  }
+  
+  go()